+++ react-app/src/components/CreateDeck/CreateDeck.js
@@ -1,10 +1,10 @@
 import React from 'react'
 
-import CheckOutDeck from './CheckoutDeck';
-import PublicDecks from './PublicDecks';
+// import CheckOutDeck from '../Library/CheckoutDeck';
+// import PublicDecks from '../Library/PublicDecks';
 import YourCards from './YourCards'
-import SearchBar from './SearchBar.js';
-import MinorAction from './MinorAction.js';
+import SearchBar from '../FullPageDiv/SearchBar.js';
+import MinorAction from '../FullPageDiv/MinorAction.js';
 import AddCardToDeck from './AddCardToDeck';
 
 const CreateDeck = () => {
