+++ react-app/src/components/FullPageDiv/Banner.js
@@ -1,5 +1,7 @@
 import React from 'react'
 
+import './Banner.css'
+
 const Banner = () => {
     return (
         <div class='bannerDiv'>
