+++ react-app/src/components/FullPageDiv/MinorAction.css
@@ -1,0 +1,46 @@
+
+.minorActionDiv {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	height: 12vh;
+	background-color: whitesmoke;
+	border-top: 2px solid rgba(5, 13, 133, 0.9);
+}
+
+.createButton {
+	height: 40px;
+	width: 80px;
+	background-color: rgba(5, 13, 133, 0.9);
+	color: whitesmoke;
+	border-radius: 15%;
+	margin-left: 20px;
+}
+
+.folderButton {
+	height: 40px;
+	width: 80px;
+	background-color: whitesmoke;
+	color: black;
+	border-radius: 15%;
+	margin-left: 20px;
+	color: rgba(5, 13, 133, 0.9);
+	border: 2px solid rgba(5, 13, 133, 0.9);
+}
+
+.selectButton {
+	margin-left: 10px;
+	height: 40px;
+	width: 80px;
+	background-color: seashell;
+	color: black;
+	border-radius: 5%;
+	color: rgba(5, 13, 133, 0.9);
+	border: 2px solid rgba(5, 13, 133, 0.9);
+}
+
+.sortDiv {
+	margin-right: 15px;
+	color: rgba(5, 13, 133, 0.9);
+	font-size: 20px;
+}
