+++ react-app/src/components/FullPageDiv/SearchBar.js
@@ -1,6 +1,6 @@
 import React, {useState} from "react";
 import { useDispatch, useSelector } from "react-redux";
-import * as deckActions from "../store/decks"
+import * as deckActions from "../../store/decks"
 
 const SearchBar = () => {
 	const [term, setTerm] = useState('')
