+++ react-app/src/components/FullPageDiv/SideBar.js
@@ -1,5 +1,7 @@
 import React from 'react'
 
+import './SideBar.css'
+
 const SideBar = () => {
     return (
         <div class='sideBarDiv'>
