+++ react-app/src/components/FullPageDiv/YourDecks.js
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import './YourDecks.css'
+
 const YourDecks = () => {
     return (
         <div class='yourDecksDiv'>
